Moving the cache objects into their own directory to cleanup `pods/`
